requirements.txt
beautifulsoup4==4.13.4
black==25.1.0
loguru==0.7.3
lxml==5.4.0
pandas==2.2.3
pytest==8.4.0
python-dotenv==1.1.0
requests==2.32.3
